+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ### Added
-- ReportingApi with attachment support [#122](https://github.com/reportportal/agent-js-jest/issues/122).
+- ReportingApi with attachment support. Resolves [#122](https://github.com/reportportal/agent-js-jest/issues/122).
 ### Changed
 - **Breaking change** Drop support of Node.js 10. The version [5.0.8](https://github.com/reportportal/agent-js-jest/releases/tag/v5.0.8) is the latest that supports it.
-- `@reportportal/client-javascript` bumped to version `5.1.3`.
+- `@reportportal/client-javascript` bumped to version `5.1.4`.
+### Security
+- Updated versions of vulnerable packages (braces).
+### Deprecated
+- Node.js 12 usage. This version is the latest that supports Node.js 12.
 
 ## [5.0.8] - 2024-01-19
 ### Deprecated