+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Added
-- ReportingApi with attachment support. Resolves [#122](https://github.com/reportportal/agent-js-jest/issues/122).
+- `ReportingApi` with `attachment` method support. Resolves [#122](https://github.com/reportportal/agent-js-jest/issues/122).
+- `extendTestDescriptionWithLastError` config option to extend test description with last error message. Resolves [#136](https://github.com/reportportal/agent-js-jest/issues/136). Thanks to [artsiomBandarenka](https://github.com/artsiomBandarenka).
 ### Changed
 - **Breaking change** Drop support of Node.js 10. The version [5.0.8](https://github.com/reportportal/agent-js-jest/releases/tag/v5.0.8) is the latest that supports it.
 - `@reportportal/client-javascript` bumped to version `5.1.4`.
@@ -7,6 +8,7 @@
 - Updated versions of vulnerable packages (braces).
 ### Deprecated
 - Node.js 12 usage. This version is the latest that supports Node.js 12.
+- `token` config option. Use `apiKey` instead.
 
 ## [5.0.8] - 2024-01-19
 ### Deprecated